Created useTasks hook + Task component
The state of the tasks has been moved to its own Context, this was
done so that future implmentation of the task will be centralized.
For example, creating a task would simply just call the useTasks
dispatch function.

The TaskBoard component now acts as a filter for the different
tasks state. Any future addition of the tasks state will simply have
to be modified here.