Created Log Component
Several Things had to be changed to the Task and TaskInfo
component so that the Logs would be as uptodate and mutaable.
When a Log is started the active log is also included to in the logs
and can be deleted later on.

The Time component has been moved to it's own component becuase
it is also used in the Log component.

A utility module has been created for miscellaneous functions that
is used for several components.